kiosk/requirements.txt
Flask
psycopg[binary,pool]
requests
Flask-DebugToolbar
flask_apscheduler
gunicorn
python-dotenv
